refactor: 패키지 방식 변경
기존 모델, 뷰, 프레젠터끼리 모아놓았던 방식에서 액티비티 별로 패키지를 갖도록 하고 각 패키지가 모델, 뷰, 프레젠터를 갖고 있도록 변경함
