+++ app/src/androidTest/java/woowacourse/movie/java/MovieReservationTest.kt
@@ -12,7 +12,7 @@ import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
 import woowacourse.movie.R
-import woowacourse.movie.reservation.MovieReservationActivity
+import woowacourse.movie.reservation.view.MovieReservationActivity
 
 @RunWith(AndroidJUnit4::class)
 class MovieReservationTest {
