+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:theme="@style/Theme.Movie"
         tools:targetApi="31">
         <activity
-            android:name=".list.MovieListActivity"
+            android:name=".list.view.MovieListActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -22,10 +22,10 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".reservation.MovieReservationActivity"
+            android:name=".reservation.view.MovieReservationActivity"
             android:exported="false" />
         <activity
-            android:name=".ticket.MovieTicketActivity"
+            android:name=".ticket.view.MovieTicketActivity"
             android:exported="false" />
     </application>
