+++ app/src/main/java/woowacourse/movie/list/adapter/MovieViewHolder.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.list
+package woowacourse.movie.list.adapter
 
 import android.widget.Button
 import android.widget.ImageView
