--- app/src/main/java/woowacourse/movie/reservation/MovieReservationMovieData.kt
@@ -1,7 +1,0 @@
-package woowacourse.movie.reservation
-
-import woowacourse.movie.list.Movie
-
-object MovieReservationMovieData {
-    var movieData: Movie? = null
-}
