+++ app/src/main/java/woowacourse/movie/reservation/model/MovieReservationMovieData.kt
@@ -1,0 +1,7 @@
+package woowacourse.movie.reservation.model
+
+import woowacourse.movie.list.model.Movie
+
+object MovieReservationMovieData {
+    var movieData: Movie? = null
+}
