+++ app/src/main/java/woowacourse/movie/ticket/presenter/MovieTicketPresenter.kt
@@ -1,4 +1,7 @@
-package woowacourse.movie.ticket
+package woowacourse.movie.ticket.presenter
+
+import woowacourse.movie.ticket.contract.MovieTicketContract
+import woowacourse.movie.ticket.model.TicketDataResource
 
 class MovieTicketPresenter(
     val view: MovieTicketContract.View,
