+++ app/src/main/res/layout/activity_movie_ticket.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:gravity="center_horizontal"
-    tools:context=".ticket.MovieTicketActivity"
+    tools:context=".ticket.view.MovieTicketActivity"
     android:orientation="vertical">
     <LinearLayout
         android:background="#9160F9"
